+++ pkg/util/docker/docker_test.go
@@ -13,7 +13,6 @@ import (
 
 	"github.com/DataDog/datadog-agent/pkg/util/containers"
 	"github.com/docker/docker/api/types"
-	"github.com/docker/docker/api/types/container"
 	"github.com/docker/docker/api/types/network"
 	"github.com/stretchr/testify/assert"
 )
@@ -425,81 +424,5 @@ func TestParseContainerNetworkAddresses(t *testing.T) {
 		for _, addr := range tc.expected {
 			assert.Contains(t, networkAddresses, addr, "test %d failed: %s", i, tc.name)
 		}
-	}
-}
-
-func TestParseContainerNetworkMode(t *testing.T) {
-	tests := []struct {
-		name       string
-		hostConfig *container.HostConfig
-		want       string
-		wantErr    bool
-	}{
-		{
-			name: "default",
-			hostConfig: &container.HostConfig{
-				NetworkMode: "default",
-			},
-			want:    "default",
-			wantErr: false,
-		},
-		{
-			name: "host",
-			hostConfig: &container.HostConfig{
-				NetworkMode: "host",
-			},
-			want:    "host",
-			wantErr: false,
-		},
-		{
-			name: "bridge",
-			hostConfig: &container.HostConfig{
-				NetworkMode: "bridge",
-			},
-			want:    "bridge",
-			wantErr: false,
-		},
-		{
-			name: "none",
-			hostConfig: &container.HostConfig{
-				NetworkMode: "none",
-			},
-			want:    "none",
-			wantErr: false,
-		},
-		{
-			name: "awsvpc",
-			hostConfig: &container.HostConfig{
-				NetworkMode: "container:0a8f83f35f7d0161f29b819d9b533b57acade8d99609bba63664dd3326e4d301",
-			},
-			want:    "awsvpc",
-			wantErr: false,
-		},
-		{
-			name: "unknown",
-			hostConfig: &container.HostConfig{
-				NetworkMode: "unknown network",
-			},
-			want:    "",
-			wantErr: true,
-		},
-		{
-			name:       "nil hostConfig",
-			hostConfig: nil,
-			want:       "",
-			wantErr:    true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := parseContainerNetworkMode(tt.hostConfig)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("parseContainerNetworkMode() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if got != tt.want {
-				t.Errorf("parseContainerNetworkMode() = %v, want %v", got, tt.want)
-			}
-		})
 	}
 }
